修改工具类和常量

+++ app/src/main/java/com/wzhy/myana/general/utils/DisplayUtil.java
@@ -6,38 +6,64 @@ package com.wzhy.myana.general.utils;
  */
 
 public class DisplayUtil implements Util{
-
+    /**
+     * dp转px
+     */
     public static int dp2px(int dpValue) {
         return ((int) (dp2pxF(dpValue) + 0.5));
     }
+
+    /**
+     * px转dp
+     */
     public static int px2dp(int pxValue) {
         return ((int) (px2dpF(pxValue) + 0.5));
     }
 
+    /**
+     * sp转px
+     */
     public static int sp2px(int spValue) {
         return ((int) (sp2pxF(spValue) + 0.5));
     }
 
+    /**
+     * px转sp
+     */
     public static int px2sp(int pxValue) {
         return ((int) (px2spF(pxValue) + 0.5));
     }
 
-    public static float dp2pxF(int dpValue) {
+//**//**//**//**//**//**//**//**//**//**//**//**//**//**//**//**//**//**//**//**//**//**//**//**//
+
+    /**
+     * dp转px
+     */
+    public static float dp2pxF(float dpValue) {
         float density = RES.getDisplayMetrics().density;
         return dpValue * density;
     }
 
-    public static float px2dpF(int pxValue) {
+    /**
+     * px转dp
+     */
+    public static float px2dpF(float pxValue) {
         float density = RES.getDisplayMetrics().density;
         return pxValue / density;
     }
 
-    public static float sp2pxF(int spValue) {
+    /**
+     * sp转px
+     */
+    public static float sp2pxF(float spValue) {
         float fontScale = RES.getDisplayMetrics().scaledDensity;
         return spValue * fontScale;
     }
 
-    public static float px2spF(int pxValue) {
+    /**
+     * px转sp
+     */
+    public static float px2spF(float pxValue) {
         float fontScale = RES.getDisplayMetrics().scaledDensity;
         return pxValue / fontScale;
     }
